requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
flasgger==0.9.7.1
pytest==7.4.3
pytest-flask==1.2.0
passlib==1.7.4
markupsafe==2.1.3
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2